refactor: 그룹채팅 컴포넌트별 분리
